requirements.txt
aiofiles==0.8.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blis==0.7.6
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cymem==2.0.6
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastapi==0.68.2
ffmpeg-python==0.2.0
future==0.18.2
h11==0.13.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
langcodes==3.3.0
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
moviepy==1.0.3
murmurhash==1.0.6
nbclassic==0.3.6
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
peewee==3.14.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip-review==1.1.1
plac==1.1.3
preshed==3.0.6
prodigy==1.11.7
proglog==0.1.9
prometheus-client==0.13.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
starlette==0.14.2
terminado==0.13.2
testpath==0.6.0
thinc==8.0.13
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.13.4
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
zipp==3.7.0
